Refactor: token.js 코드 리팩토링
앱에서 토큰을 받아오는 getServerToken 함수 및 token.js 내부의 코드 정리
파일의 전역 스코프에 있는 변수를 클로저를 활용하여 자유변수 변경

Changes to be committed:
modified:   src/pages/ExerciseRoutine/ExerciseRoutine.tsx
deleted:    src/utils/.empty
deleted:    src/utils/mobile/healthyup.js
new file:   src/utils/mobile/token.js
